Added useful functions to FrameProxy
