+++ mzLib/Readers/Bruker/TimsTofReader/TimsTofFileReader.cs
@@ -105,7 +105,7 @@ namespace Readers.Bruker
             int frameIdIndex = 65365;
             long frameId = ids[frameIdIndex];
 
-            var dataArray = GetScanRawData(fileHandle, ids[frameIdIndex], (UInt32)numScans[frameIdIndex]);
+            var dataArray = FrameProxy.GetScanRawData(fileHandle, ids[frameIdIndex], (UInt32)numScans[frameIdIndex]);
             FrameProxy test = new FrameProxy(dataArray, numScans[frameIdIndex], fileHandle);
 
             var max = dataArray.Max();
@@ -123,76 +123,17 @@ namespace Readers.Bruker
 
             placeholded2++;
 
-            var conversionTest = BdalTimsConversionFunctions.DoTransformation(fileHandle,
+            var conversionTest = TimsConversion.DoTransformation(fileHandle,
                 ids[frameIdIndex],
                 mzDubSlice,
-                BdalTimsConversionFunctions.ConversionFunctions.IndexToMz);
+                TimsConversion.ConversionFunctions.IndexToMz);
 
             placeholded2++;
 
         }
 
-        /// <summary>
-        /// This is reimplementation of the Marshal.Copy method that allows for arbitrary types
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="source"></param>
-        /// <param name="destination"></param>
-        /// <param name="startIndex"></param>
-        /// <param name="length"></param>
-        /// <exception cref="ArgumentNullException"></exception>
-        /// <exception cref="ArgumentOutOfRangeException"></exception>
-        private static unsafe void CopyToManaged<T>(IntPtr source, T[] destination, int startIndex, int length)
-        {
-            if (source == IntPtr.Zero) throw new ArgumentNullException(nameof(source));
-            if (destination is null) throw new ArgumentNullException(nameof(destination));
-            if (startIndex < 0) throw new ArgumentOutOfRangeException(nameof(startIndex));
-            if (length < 0) throw new ArgumentOutOfRangeException(nameof(length));
-
-            void* sourcePtr = (void*)source;
-            Span<T> srcSpan = new Span<T>(sourcePtr, length);
-            Span<T> destSpan = new Span<T>(destination, startIndex, length);
-
-            srcSpan.CopyTo(destSpan);
-        }
-
-        internal unsafe static uint[] GetScanRawData(UInt64 fileHandle, long frameId, UInt32 numScans)
-        {
-            int dataLength = 4096; // Default allocation ~ 16 kB
-            
-            // buffer expansion loop
-            while(true)
-            {
-                IntPtr pData = Marshal.AllocHGlobal(dataLength * Marshal.SizeOf<Int32>());
-
-                var outputLength = tims_read_scans_v2(
-                    fileHandle,
-                    frameId,
-                    scan_begin: 0,
-                    scan_end: numScans,
-                    buffer: pData,
-                    length: (uint)(dataLength * 4));
-
-                if (4 * dataLength > outputLength)
-                {
-                    var dataArray = new uint[dataLength];
-                    CopyToManaged(pData, dataArray, 0, dataLength);
-                    Marshal.FreeHGlobal(pData);
-
-                    return dataArray;
-                }
-
-                if(outputLength > 16777216 ) // Arbitrary 16 mb frame limit
-                {
-                    throw new Exception("Maximum frame size exceeded");
-                }
-
-                // Increase buffer size if necessary
-                dataLength = ((int)outputLength / 4) + 1;
-
-                Marshal.FreeHGlobal(pData);
-            }
-        }
+        
+        
 
         internal unsafe static void ParseScan(uint[] rawData)
         {
@@ -212,28 +153,7 @@ namespace Readers.Bruker
         public static extern UInt64 tims_open
               (byte[] analysis_directory_name_utf8, UInt32 use_recalibrated_state);
 
-        /// <summary>
-        /// Read a range of scans from a single frame.
-        ///
-        /// Output layout: (N = scan_end - scan_begin = number of requested scans)
-        ///   N x uint32_t: number of peaks in each of the N requested scans
-        ///   N x (two uint32_t arrays: first indices, then intensities)
-        ///
-        /// Note: different threads must not read scans from the same storage handle
-        /// concurrently.
-        /// </summary> 
-        /// <param name="handle"> Unique Handle of .d file ( returned on tims_open() )</param>
-        /// <param name="frame_id"> From .tdf SQLite: Frames.Id </param>
-        /// <param name="scan_begin"> first scan number to read (inclusive) </param>
-        /// <param name="scan_end"> Last scan number (exclusive) </param>
-        /// <param name="buffer"> Destination buffer allocated by user </param>
-        /// <param name="length"> Length of the buffer (in bytes, i.e. 4 * buffer.length) </param>
-        /// <returns> 0 on error, otherwise the number of buffer bytes necessary for the output
-        /// of this call (if this is larger than the provided buffer length, the result is not
-        /// complete). </returns>
-        [DllImport("Bruker/TimsTofReader/timsdata.dll", CallingConvention = CallingConvention.Cdecl)]
-        unsafe static extern UInt32 tims_read_scans_v2
-              (UInt64 handle, Int64 frame_id, UInt32 scan_begin, UInt32 scan_end, IntPtr buffer, UInt32 length);
+        
 
         #endregion Bruker Dll Functions
